+++ packages/realm_dart/lib/src/native/map_handle.dart
@@ -3,9 +3,9 @@
 
 import 'dart:ffi';
 
-import 'package:ffi/ffi.dart';
+import 'ffi.dart';
 
-import '../realm_dart.dart'; // TODO: remove this import
+import '../realm_dart.dart';
 import 'convert_native.dart';
 import 'error_handling.dart';
 import 'map_changes_handle.dart';
@@ -13,7 +13,7 @@ import 'notification_token_handle.dart';
 import 'object_handle.dart';
 import 'query_handle.dart';
 import 'realm_bindings.dart';
-import 'realm_core.dart'; // TODO: remove this import
+// TODO: remove this import
 import 'realm_handle.dart';
 import 'realm_library.dart';
 import 'results_handle.dart';
