+++ packages/realm_dart/lib/src/native/object_handle.dart
@@ -3,9 +3,9 @@
 
 import 'dart:ffi';
 
-import 'package:ffi/ffi.dart';
+import 'ffi.dart';
 
-import '../realm_dart.dart'; // TODO: Remove this import
+import '../realm_dart.dart';
 import 'convert_native.dart';
 import 'error_handling.dart';
 import 'handle_base.dart';
@@ -13,7 +13,6 @@ import 'list_handle.dart';
 import 'map_handle.dart';
 import 'notification_token_handle.dart';
 import 'realm_bindings.dart';
-import 'realm_core.dart'; // TODO: Remove this import
 import 'realm_handle.dart';
 import 'realm_library.dart';
 import 'results_handle.dart';
