+++ packages/realm_dart/lib/src/native/schema_handle.dart
@@ -3,7 +3,7 @@
 
 import 'dart:ffi';
 
-import 'package:ffi/ffi.dart';
+import 'ffi.dart';
 import 'package:realm_common/realm_common.dart';
 
 import '../configuration.dart';
@@ -11,7 +11,6 @@ import 'handle_base.dart';
 import 'realm_bindings.dart';
 import 'realm_library.dart';
 import 'to_native.dart';
-
 
 class SchemaHandle extends HandleBase<realm_schema> {
   SchemaHandle(Pointer<realm_schema> pointer) : super(pointer, 24);
