+++ packages/realm_dart/lib/src/native/set_handle.dart
@@ -3,15 +3,14 @@
 
 import 'dart:ffi';
 
-import 'package:ffi/ffi.dart';
+import 'ffi.dart';
 
-import '../realm_dart.dart'; // TODO: remove this import
+import '../realm_dart.dart';
 import 'convert_native.dart';
 import 'error_handling.dart';
 import 'notification_token_handle.dart';
 import 'query_handle.dart';
 import 'realm_bindings.dart';
-import 'realm_core.dart'; // TODO: Remove this import
 import 'realm_handle.dart';
 import 'realm_library.dart';
 import 'results_handle.dart';
