+++ packages/realm_dart/test/app_test.dart
@@ -322,7 +322,7 @@ void main() {
     final text = isFlutterPlatform ? "realm_tests" : "realm_dart";
     const salt = [82, 101, 97, 108, 109, 32, 105, 115, 32, 103, 114, 101, 97, 116];
     final expected = base64Encode(sha256.convert([...salt, ...utf8.encode(text)]).bytes);
-    expect(getBundleId(), expected);
+    expect(realmCore.getBundleId(), expected);
   });
 
   test('app.getById without apps returns null', () {
